refactor(common): organize data dto entity packages
